+++ api/tests/test_api.py
@@ -5,16 +5,20 @@ from django.urls import reverse
 
 from posts.models import Post, ContentType
 
-from .constants import POST_DATA, POST_IMG_DATA
+from posts.tests.constants import POST_DATA
+from .constants import POST_IMG_DATA
+
+TEST_USERNAME = 'bob'
+TEST_PASSWORD = 'password'
 
 
 class AuthorTests(TestCase):
     def setUp(self) -> None:
         self.client = Client()
-        get_user_model().objects.create_user(username='bob', password='password')
+        get_user_model().objects.create_user(username=TEST_USERNAME, password=TEST_PASSWORD)
 
     def test_authors(self):
-        self.client.login(username='bob', password='password')
+        self.client.login(username=TEST_USERNAME, password=TEST_PASSWORD)
         res = self.client.get('/api/v1/authors/')
         self.assertEqual(res.status_code, 200)
         body = json.loads(res.content.decode('utf-8'))
@@ -26,9 +30,19 @@ class AuthorTests(TestCase):
         self.assertEqual(res.status_code, 403)
 
     def test_create_author(self):
-        self.client.login(username='bob', password='password')
-        res = self.client.post('/api/v1/authors/', {'username': 'alice', 'password': 'password'})
+        self.client.login(username=TEST_USERNAME, password=TEST_PASSWORD)
+        res = self.client.post('/api/v1/authors/', {'username': 'alice', 'password': 'some_password'})
         self.assertEqual(res.status_code, 405)
+
+    def test_allow_api_users(self):
+        api_user_username = 'api_user'
+        api_user = get_user_model().objects.create_user(username=api_user_username, password=TEST_PASSWORD)
+        api_user.is_api_user = True
+        api_user.save()
+
+        self.client.login(username=api_user_username, password=TEST_PASSWORD)
+        res = self.client.get('/api/v1/authors/')
+        self.assertEqual(res.status_code, 200)
 
 
 class ImageTests(TestCase):
